modules/commerce-services/commercefacades/src/de/hybris/platform/commercefacades/order/converters/populator/PaymentModePopulator.java
/*
 * Copyright (c) 2019 SAP SE or an SAP affiliate company. All rights reserved.
 */
package de.hybris.platform.commercefacades.order.converters.populator;

import de.hybris.platform.commercefacades.order.data.PaymentModeData;
import de.hybris.platform.converters.Populator;
import de.hybris.platform.core.model.order.payment.PaymentModeModel;
import de.hybris.platform.servicelayer.dto.converter.ConversionException;

import org.springframework.util.Assert;



public class PaymentModePopulator implements Populator<PaymentModeModel, PaymentModeData>
{

	@Override
	public void populate(final PaymentModeModel source, final PaymentModeData target) throws ConversionException
	{
		Assert.notNull(source, "Parameter source cannot be null.");
		Assert.notNull(target, "Parameter target cannot be null.");

		target.setCode(source.getCode());
		target.setName(source.getName());
		target.setDescription(source.getDescription());
	}

}
